pyproject.toml
[project]
name = "fyp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.5",
    "pygame-ce>=2.5.5",
    "pymaze>=1.2.3",
]
